[SERVICE-528] Move validator code
